pyproject.toml
[project]
name = "tribu-ia-purple-agent"
version = "0.1.0"
description = "A purple agent for code generation benchmarks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "litellm>=1.20.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
]
